refactor: 유니티 통신 인터페이스 객체화 및 좌표 일관성 향상 (#38)
- Unity 브리지 메서드의 파라미터를 Primitive 나열 방식에서 `TextElement`, `RgbaColor` 등 객체 전달 방식으로 변경하여 캡슐화 강화
- 객체 배치의 일관성과 정밀도를 높이기 위해 모든 좌표(`posX`, `posY`, `posZ`) 타입을 `Int`에서 `Float`으로 변경
- `RgbaColor` -> `ColorDto` 변환 확장 함수(`toDto`)를 도입하여 색상 변환 로직의 중복 제거
- Unity 메시지 핸들러 메서드명 `...FromAndroid` 접미사를 추가하도록 수정
- 신규 오브젝트 생성 시 초기 좌표 계산 로직을 `Float` 타입에 맞춰 수정

+++ core/data/src/main/java/com/itschristmas/data/repositoryImpl/CardElementRepositoryImpl.kt
@@ -59,12 +59,12 @@ class CardElementRepositoryImpl @Inject constructor(
             cardElementDao.getTextElementsByCardId(cardId).map { it.toDomain() }
         }
 
-    override suspend fun updateElementTransform(elementId: Long, posX: Int, posY: Int, scale: Int): Result<Int> =
+    override suspend fun updateElementTransform(elementId: Long, posX: Float, posY: Float, scale: Int): Result<Int> =
         ioCatching {
             cardElementDao.updateElementTransform(elementId, posX, posY, scale)
         }
 
-    override suspend fun updateTextElement(elementId: Long, textAttributes: TextAttributes, posX: Int, posY: Int): Result<Int> =
+    override suspend fun updateTextElement(elementId: Long, textAttributes: TextAttributes, posX: Float, posY: Float): Result<Int> =
         ioCatching {
             cardElementDao.updateTextElement(
                 elementId = elementId,
